// src/app/models/product.interface.ts
export interface IProduct {
    id: number;
    code: string;
    barcode: string;
    count: number;
    price: number;
    originalPrice: number;
    costPrice: number;
    isSale: boolean;
    title: string;
    createdAt: string;
    modifiedAt: string;
    expiredAt: string;
    imageUrlsJson: string;
    avatarUrl: string;
    unit: string;
    foreignCurrency: string;
    costPriceForeign: number;
    onlineId: number;
    isOption: boolean;
    isCombo: boolean;
    isPublic: boolean;
    isService: boolean;
    status: number;
    itemsJson: string;
    items: any[];
    showOnWeb: boolean;
    isHotProduct: boolean;
    isNewProduct: boolean;
    showPriceOnWeb: boolean;
    description: string;
    isSelected: boolean;
    storeQuantity: number;
    units: any[];
    unitsJson: string;
    fromUnit: string;
    staffId: number;
    isMaterial: boolean;
    materialsJson: string;
    materials: any[];
    autoMaterials: boolean;
}